(refact) cleans up server side code
