+++ database-mongo/index.js
@@ -14,58 +14,3 @@ db.once('open', function() {
 });
 
 module.exports = db;
-
-  //TESTING FUNCTIONALITY
-
-  // var mike = new User({
-  //   name: 'mike',
-  //   isOwner: false,
-  //   isAdmin: true,
-  //   join: new Date,
-  //   meta: {
-  //     reviews: 10,
-  //     friends: 10
-  //   }
-  // })
-
-  // var speedyReview = new Review({
-  //   content: 'awesome wine, great value',
-  //   product: 'speedy creek zin',
-  //   rating: 5,
-  //   user_id: 3
-  // });
-
-  // var speedyProduct = new Product({
-  //   name: 'Speedy Creek',
-  //   year: 2000,
-  //   type: 'red',
-  //   origin: 'napa',
-  //   varietal: 'zin',
-  //   price: 25,
-  //   owner_id: 2098
-  // })
-
-  // mike.save(function (err, mike) {
-  //   if(err) return console.error(err);
-  //   console.log('mike:', mike);
-  // })
-
-  // speedyReview.save(function (err, speedyReview) {
-  //   if(err) return console.error(err);
-  //   console.log('speedyReview', speedyReview);
-  // })
-
-  // speedyProduct.save(function (err, speedyProduct) {
-  //   if(err) return console.error(err);
-  //   console.log('speedyProduct', speedyProduct);
-  // })
-
-  // User.find(function (err, users) {
-  //   if(err) console.error(err);
-  //   console.log('User:', User);
-  // })
-
-  // User.find({ name: /^mike/ }, function(err, results) {
-  //   if(err) console.error(err);
-  //   console.log('find results for mike: ', results);
-  // })
