+++ server/index.js
@@ -1,6 +1,5 @@
 var express = require('express');
 var bodyParser = require('body-parser');
-var db = require('./utilities/dbUtils.js');
 var wineApiUtils = require('./utilities/wineApiUtils.js');
 var cors = require('cors');
 var passport = require('passport');
@@ -9,50 +8,19 @@ var fb = require('./utilities/config.js');
 var ppUtils = require('./utilities/passportUtils.js');
 var User = require('../database-mongo/models/User.js');
 var path = require('path');
+var Product = require('./models/product.js');
+var User = require('./models/user.js');
+var Review = require('./models/review.js');
 
 //initializing Passport with FB OAuth
 passport.use(new FacebookStrategy({
     clientID: fb.fbAppId,
     clientSecret: fb.secret,
     callbackURL: 'http://localhost:3000/login/facebook/callback',
-  },
-  function(accessToken, refreshToken, profile, done) {
-    
-    User.find({name: profile.displayName}, function(err, user) {
-      if (err) {
-        console.error(err);
-        return done(err);
-      } 
-      if (user.length < 1) {
-        User.create({
-          name: profile.displayName,
-          joined: new Date(),
-          accessToken: accessToken,
-          meta: {
-            reviews: 0,
-            friends: 0,
-          }, 
-        }, function(err, user) {
-          if (err) {
-            return done(err);
-          } else {
-            return done(null, user);
-          }
-        });
-      } else {
-        done(null, user);
-      }
-    });
-  }
-));
+  }, User.passportAuth));
 
-
-
-//INSTANTIATE APP
 var app = express();
 
-
-//MIDDLEWARE
 app.use(bodyParser.json());
 app.use(cors());
 app.use(require('cookie-parser')());
@@ -61,8 +29,6 @@ app.use(require('express-session')({secret: 'keyboard cat', resave: true, saveUn
 app.use(passport.initialize());
 app.use(passport.session());
 
-
-// AUTHENTICATION
 app.get('/', ppUtils.isLoggedIn, (req, res) => {
   res.sendFile(path.join(__dirname, '/../react-client/dist'));
 });
@@ -88,15 +54,11 @@ app.get('/login/facebook',
 
 app.get('/login/facebook/callback', 
   passport.authenticate('facebook', { failureRedirect: '/login',
-                                      // session: false,
-                                      successRedirect: '/'
-                                    })
+    // session: false,
+    successRedirect: '/'
+  })
 );
 
-
-//SETTING UP ALL THE ROUTES FOR THE CLIENT REQUEST
-
-// initializes the red and white wines on the client
 app.get('/init', function(req, res) {
   var wines = {
     top10Reds: [],
@@ -104,12 +66,12 @@ app.get('/init', function(req, res) {
     topRated: [],
   };
 
-  db.top10Reds(function(error, topReds) {
+  Product.top10Reds(function(error, topReds) {
     if (error) {
       res.send(error);
     } else {
       wines.top10Reds = topReds;
-      db.top10Whites(function(error, topWhites) {
+      Product.top10Whites(function(error, topWhites) {
         if(error){
           res.send(error)
         } else {
@@ -139,9 +101,9 @@ app.get('/getWines', function(req, res) {
   var price = 10;
 
   wineApiUtils.topRed(price, function(error, results) {
-    db.storeWines(results.Products.List, function() {
+    Product.storeWines(results.Products.List, function() {
       wineApiUtils.topWhite(price, function(error, results) {
-        db.storeWines(results.Products.List, function() {
+        Product.storeWines(results.Products.List, function() {
           res.send('storage of red and white wines complete!');
         });
       });
@@ -152,38 +114,33 @@ app.get('/getWines', function(req, res) {
 //This route invokes wine.com api.
 // app.get('/catalog', wineApiUtils.apiRequest);
 
-//POST request for search
 app.post('/search', function(req, res) {
   var query = req.body.search;
   var price = req.body.price || 10;
 
-  console.log('search!!!!!!!!!!!!!!!:', query, price);
-
-  db.searchWines(query, price, function(error, results) {
+  Product.searchWines(query, price, function(error, results) {
     if(error){
       console.log('Error from server API request', error);
       res.sendStatus(404).send('not found')
     } else {
-      // console.log('results from search post API', results);
       res.send(results);
     }
   })
 
 });
 
-//POST request for signup
 app.post('/signup', function(req, res) {
   var user = req.body.username;
   var pass = req.body.password;
 
   //check for valid username, i.e. currently not in use
-  db.checkuserName(user, function(error, valid, results){
+  User.checkuserName(user, function(error, valid, results){
     if(error){
       res.send('error inside checkuserName index');
     } else if (!valid) {
       res.send('duplicate username')
     } else if (valid) {
-      db.addUser(user, pass, function(error, success, results){
+      User.addUser(user, pass, function(error, success, results){
         if(error) {
           res.send('error inside addUser index.js');
         } else if (success) {
@@ -197,7 +154,7 @@ app.post('/signup', function(req, res) {
 app.post('/users/username/', function(req, res) {
   var username = req.body.username;
 
-  db.checkuserName(username, function(error, valid, results){
+  User.checkuserName(username, function(error, valid, results){
     if(error){
       console.error(error)
       res.send(error);
@@ -210,12 +167,11 @@ app.post('/users/username/', function(req, res) {
 
 app.options('/users/username/');
 
-//POST request for login
 app.post('/login', function(req, res) {
   var username = req.body.username;
   var password = req.body.password;
 
-  db.validateUser(username, password, function(error, results) {
+  User.validateUser(username, password, function(error, results) {
     if(error){
       console.log(error)
     } else {
@@ -224,7 +180,6 @@ app.post('/login', function(req, res) {
   })
 });
 
-//POST request for review
 app.post('/review', function(req, res) {
   var content = req.body.review;
   var rating = req.body.rating;
@@ -238,7 +193,7 @@ app.post('/review', function(req, res) {
     username: username,
     product_id: product_id
   }
-  db.addReview(review, function(error, results){
+  Review.addReview(review, function(error, results){
     if(error){
       console.log('error inside dbUtils addReview', error);
     } else {
@@ -248,152 +203,23 @@ app.post('/review', function(req, res) {
   })
 });
 
-//GET ALL REVIEWS FOR RESPECTIVE PRODUCT
 app.post('/reviews', function(req, res) {
   var product_id = req.body.product_id;
   console.log('product inside reviews GET all', product_id);
 
-  db.getReviews(product_id, function(error, reviews){
+  Review.getReviews(product_id, function(error, reviews){
     if(error){
-      console.log('error!!!!!!!!!!!!')
+      console.log('error in post-reviews')
       res.send(error)
     } else {
-      console.log('success!!!!!!!!!!');
-      console.log('reviews', reviews)
       res.send(reviews)
     }
   })
 });
 
-// var pricesArray = [0, 10, 20, 30, 40, 50];
-// var winesArray = ['red', 'white', 'rose', 'Cabernet Franc', 'cabernet', 'fran', 'cabernet sauvignon', 'gamay', 'grenache', 'garnacha', 'malbec', 'merlot', 'mourvedre', 'mataro', 'nebbiolo', 'pinot', 'pinot noir', 'sango', 'sangiovese', 'shiraz', 'syrah', 'zin', 'zinfandel', 'chenin', 'chenin blanc', 'blanc', 'Gewurztraminer', 'marsanne', 'muscat', 'pinot blanc', 'blanc', 'pinot gris', 'pinot grigio', 'riesling', 'roussanne', 'sauvignon blanc', 'fume blanc', 'semillon', 'viognier', 'gruner veltliner', 'brut']
-
-// winesArray.forEach(function(wine){
-
-//   pricesArray.forEach(function(price){
-
-//     wineApiUtils.forcedRequest(price, wine, function(error, results) {
-//       if(error){
-//         console.log('error inside forcedRequest', error);
-//       } else {
-
-//         resBody = JSON.parse(results.body);
-//         var wines = resBody.Products.List;
-//         wines.forEach(function(wine){
-
-//           var query = {
-//             name: wine.Name,
-//             year: wine.Vintage + '',
-//             type: wine.Varietal.Name,
-//             redORwhite: wine.Varietal.WineType.Name,
-//             origin: wine.Appellation.Name,
-//             region: wine.Appellation.Region.Name,
-//             priceMin: wine.PriceMin,
-//             priceMax: wine.PriceMax,
-//             apiRating: wine.Ratings.HighestScore
-//           }
-
-//           db.addWine(query, function(error, results){
-//             if(error){
-//               console.error(error)
-//             } else {
-//               console.log('results from addwine db', results)
-//             }
-//           })
-//         })
-//       }
-//     })
-//   })
-// })
 
 var port = process.env.PORT || 3000;
 
 app.listen(port, function() {
   console.log('listening to port ' + port);
 });
-
-
-
-
-
-      // var query = {
-      //   name: wine.Name,
-      //   year: wine.Vintage + '',
-      //   type: wine.Varietal.Name,
-      //   redORwhite: wine.Varietal.WineType.Name,
-      //   origin: wine.Appellation.Name,
-      //   region: wine.Appellation.Region.Name,
-      //   priceMin: wine.PriceMin,
-      //   priceMax: wine.PriceMax,
-      //   apiRating: wine.Ratings.HighestScore
-      // }
-
-
-      // db.addWine(query, function(error, results){
-      //   if(error){
-      //     console.error(error)
-      //   } else {
-      //     console.log('results from addwine db', results)
-      //   }
-      // })
-// app.listen(3000);
-// httpsServer.listen(3000)
-
-
-
-      // //to see how the object is structured
-
-
-//     })
-
-//   }
-// })
-      // console.log('VARIETAL', wine.Varietal)
-      // console.log('VINEYARD', wine.Vineyard)
-      // console.log('VINTAGES', wine.Vintages)
-      // console.log('GEOLOCATION',wine.GeoLocation)
-      // console.log('APPELLATION', wine.Appellation)
-
-      // //where the information we want is located
-      // console.log('NAME', wine.Name) //possibly change
-      // console.log('YEAR', wine.Vintage)
-      // console.log('TYPE', wine.Varietal.Name)
-      // console.log('REDORWHITE', wine.Varietal.WineType.Name)
-      // console.log('ORIGIN', wine.Appellation.Name)
-      // console.log('REGION', wine.Appellation.Region.Name)
-      // console.log('PRICE(MIN-MAX)', wine.PriceMin + '-' + wine.PriceMax)
-      // console.log('APIRATING', wine.Ratings.HighestScore)
-// []
-
-//In terminal
-//export PORT=3000
-//nodemon index.js or npm server-dev
-
-
-// example data format for wine product from API request
-// { Id: 167706,
-// Name: 'CVNE Crianza 2013',
-// Url: 'http://www.wine.com/v6/CVNE-Crianza-2013/wine/167706/Detail.aspx',
-// Appellation: [Object],
-// Labels: [Object],
-// Type: 'Wine',
-// Varietal: [Object],
-// Vineyard: [Object],
-// Vintage: '2013',
-// Community: [Object],
-// Description: '',
-// GeoLocation: [Object],
-// PriceMax: 15.99,
-// PriceMin: 12.99,
-// PriceRetail: 15,
-// ProductAttributes: [Object],
-// Ratings: [Object],
-// Retail: null,
-// Vintages: [Object] }
-// app.options('*', function(req, res) {
-//   res.send('awesome')
-// })
-
-// app.get('/.well-known/acme-challenge/:content', function(req, res) {
-//   res.send('PW-phkjKRkbF7Oeg1ed8b_EuymzQGBI0leXIYDq5Hyc')
-// })
